Update main
fixed the bug where two messages were being printed every broad update on 5th updates. This was done with the total_bypasser variable which now collects the total before the total is reset and passes through the loop again on % 5 != 0, the glitch was occuring because the total was being incremented by 1 before it attempted to go through the second if. The must be another solution but this works for now.